+++ elastos/crates/elastos-common/src/lib.rs
@@ -5,7 +5,6 @@ pub mod collaboration_protocol;
 mod error;
 pub mod localhost;
 mod manifest;
-pub mod protected_content;
 pub mod timestamp;
 mod types;